backend/requirements.txt
# Основные зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.0.0
loguru==0.7.2
websockets==12.0

# HTTP клиенты
httpx==0.28.0  # Обновлено для совместимости
aiohttp==3.9.1
requests==2.32.5

# Дополнительные
APScheduler==3.10.4
python-telegram-bot==22.5
phonenumbers==9.0.22
redis==5.0.1
python-multipart==0.0.6
PyYAML==6.0.1
pillow==12.0.0

# Вспомогательные
aiofiles==23.2.1
aiosqlite==0.22.1
tzlocal==5.3.1
pytz==2025.2
tzdata==2025.2